model remote changed!

+++ data/src/main/java/com/jnfran92/data/crypto/model/crypto/Price.kt
@@ -3,7 +3,7 @@ package com.jnfran92.data.crypto.model.crypto
 data class Price(
     var price: Double,
     var marketCap: Double,
-    var volume24h: Long,
+    var volume24h: Double,
     var percentChange1h: Double,
     var percentChange24h: Double,
     var percentChange7d: Double,
